requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
openpyxl==3.1.2
pandas>=2.2.3
bcrypt==4.1.2
email-validator>=2.1.1
Pillow>=10.4.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
nltk>=3.8
textblob>=0.17.1
openai>=1.0.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0